backend/requirements-min.txt
# Web Framework
fastapi==0.128.0
uvicorn==0.40.0
starlette==0.50.0

# Database
SQLAlchemy==2.0.45
psycopg2-binary==2.9.11
alembic==1.17.2

# Authentication & Security
bcrypt==4.0.1
python-jose==3.5.0
passlib==1.7.4
cryptography==46.0.3

# Data Validation
pydantic==2.12.5
pydantic_core==2.41.5
email-validator==2.3.0

# Utilities
python-dotenv==1.2.1
requests==2.32.5
python-multipart==0.0.21

# Async
httpx==0.28.1
httpcore==1.0.9
anyio==4.12.0
h11==0.16.0

# Image Processing & OCR
pillow==12.0.0
opencv-python-headless==4.12.0.88
easyocr==1.7.2

# Data Science & ML
numpy==2.2.6
pandas==2.3.3
scikit-learn==1.7.2
scipy==1.15.3
joblib==1.5.3
torch==2.9.1
torchvision==0.24.1

# Optional: Monitoring (si vous en avez besoin)
# opentelemetry-api==1.39.1
# opentelemetry-sdk==1.39.1